+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+* Replace PhantomJSControllerContext with PhantomJSControllerExtension - does
+  the same job, but as an extension.
+* Add a set of useful extensions for bootstrapping, saving failing pages,
+  providing kohana application dependencies into contexts that need them and 
+  co-ordinating and injecting test data factories and repositories (which need
+  to be actually implemented separately in the project for now). 
 * Require Behat version 3
 
 ## 0.1.0 (2018-04-25)